chore(bitcoind): reorganize code and tests
