+++ components/bitcoind/src/block_pool/tests/helpers/bitcoin_shapes.rs
