--- components/bitcoind/src/observer/mod.rs
@@ -1,1 +1,0 @@
-pub mod zmq;
