+++ components/bitcoind/src/pipeline/rpc/pipeline.rs
@@ -14,21 +14,20 @@ use reqwest::Client;
 use tokio::task::JoinSet;
 
 use crate::{
-    indexer::{
-        bitcoin::{
-            cursor::BlockBytesCursor, parse_downloaded_block, standardize_bitcoin_block,
-            try_download_block_bytes_with_retry,
+    pipeline::{
+        rpc::{
+            parse_downloaded_block, standardize_bitcoin_block, try_download_block_bytes_with_retry,
         },
         wait_for_thread_finish, BlockProcessor, BlockProcessorCommand,
     },
     try_debug, try_info,
-    types::BitcoinNetwork,
+    types::{BitcoinNetwork, BlockBytesCursor},
     utils::Context,
 };
 
 /// Downloads historical blocks from bitcoind's RPC interface and pushes them to a [BlockProcessor] so they can be indexed
 /// or ingested as needed.
-pub async fn start_block_download_pipeline(
+pub(crate) async fn start_block_download_pipeline(
     config: &Config,
     rpc_client: &Client,
     block_heights: Vec<u64>,
