+++ components/bitcoind/src/types/mod.rs
@@ -1,13 +1,8 @@
 pub mod bitcoin;
+mod cursor;
 mod ordinals;
-mod processors;
 mod rosetta;
 
+pub use cursor::*;
 pub use ordinals::*;
-pub use processors::*;
 pub use rosetta::*;
-
-#[derive(Clone, Debug)]
-pub enum Chain {
-    Bitcoin,
-}
